refactor: 패키지 명 application으로 수정
