+++ meongcare/src/main/java/com/meongcare/domain/dog/application/DogService.java
@@ -1,4 +1,4 @@
-package com.meongcare.domain.dog.service;
+package com.meongcare.domain.dog.application;
 
 import com.meongcare.domain.auth.domain.entity.Member;
 import com.meongcare.domain.auth.domain.repository.MemberRepository;
