+++ meongcare/src/main/java/com/meongcare/domain/dog/presentation/controller/DogController.java
@@ -5,8 +5,7 @@ import com.meongcare.domain.dog.presentation.dto.request.PutDogRequestDto;
 import com.meongcare.domain.dog.presentation.dto.request.SaveDogRequestDto;
 import com.meongcare.domain.dog.presentation.dto.response.GetDogResponseDto;
 import com.meongcare.domain.dog.presentation.dto.response.GetDogsResponseDto;
-import com.meongcare.domain.dog.service.DogService;
-import com.meongcare.domain.medicalrecord.presentation.dto.request.SaveMedicalRecordRequestDto;
+import com.meongcare.domain.dog.application.DogService;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
 import io.swagger.v3.oas.annotations.enums.ParameterIn;
